COMPONENTS: Refactored to new location in /app/components
