+++ src/app/components/db-name/db-name.component.html
@@ -1,0 +1,1 @@
+<a [routerLink]="['/pages/instance', value]">{{ value }}</a>
