+++ src/app/components/db-name/db-name.component.ts
