+++ src/app/pages/home/home.html
@@ -1,6 +1,6 @@
 <h1>Home</h1>
 
 <div style="padding-top: 30px">
-    <home-instances></home-instances>    
-    <home-jobs></home-jobs>
+    <instances></instances>    
+    <jobs></jobs>
 </div>
