+++ src/app/pages/pages.module.ts
@@ -4,10 +4,11 @@ import { routing }       from './pages.routing';
 import { NgaModule } from '../theme/nga.module';
 import { AppTranslationModule } from '../app.translation.module';
 import { Pages } from './pages.component';
+import { CernToolbarComponent } from '../components/cern-toolbar/cern-toolbar.component';
 
 @NgModule({
   imports: [CommonModule, AppTranslationModule, NgaModule, routing],
-  declarations: [Pages]
+  declarations: [Pages, CernToolbarComponent]
 })
 export class PagesModule {
 }
