--- src/app/theme/components/cern-toolbar/index.ts
@@ -1,1 +1,0 @@
-export * from './cern-toolbar.component';
