--- src/app/theme/components/db-name/db-name.component.html
@@ -1,1 +1,0 @@
-<a [routerLink]="['/pages/instance', value]" (click)="onClick()">{{ value }}</a>
