--- src/app/theme/components/db-name/index.ts
@@ -1,1 +1,0 @@
-export * from './db-name.component';
