+++ src/app/theme/components/index.ts
@@ -13,6 +13,3 @@ export * from './baPictureUploader';
 export * from './baCheckbox';
 export * from './baMultiCheckbox';
 export * from './baFileUploader';
-export * from './cern-toolbar';
-export * from './state-button';
-export * from './db-name';
