--- src/app/theme/components/state-button/index.ts
@@ -1,1 +1,0 @@
-export * from './state-button.component';
